Homework Vlad Coteanu

+++ CoteanuVlad/src/org/bootcamp/calculator/InsurancePolicyCalculator.java
@@ -11,7 +11,7 @@ public final class InsurancePolicyCalculator {
 
     }
 
-    public int calculate(Vehicle vehicle, Formula formula) {
+    public double calculate(Vehicle vehicle, Formula formula) {
         return formula.calculate(vehicle);
     }
 }
